pyproject.toml
[project]
name = "structflo-ner"
dynamic = ["version"]
description = "Drug discovery NER wrapper around LangExtract — zero-config entity extraction for chemistry and biology."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
dependencies = [
    "langextract>=1.1.1",
]

[project.optional-dependencies]
dataframe = ["pandas>=1.5"]
fast = ["rapidfuzz>=3.0", "PyYAML>=6.0"]

[dependency-groups]
dev = [
    "pytest>=7",
    "pytest-mock>=3",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "structflo/ner/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["structflo"]

[tool.pytest.ini_options]
testpaths = ["tests"]

# --------------------------------------------------------------------------- #
# Ruff
# --------------------------------------------------------------------------- #
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["structflo"]

[tool.ruff.format]
quote-style = "double"
